+++ pwnlib/heap/bins/tcache/tcache.py
@@ -1,4 +1,4 @@
-from pwnlib.heap.bins.bin import *
+from pwnlib.heap.bins.bin import Bins, Bin, BinEntry
 
 
 class Tcaches(Bins):
@@ -21,3 +21,16 @@ class Tcache(Bin):
 
     def __init__(self, bin_entry, malloc_chunks):
         super(Tcache, self).__init__(bin_entry, malloc_chunks)
+
+
+class TcacheEntry(BinEntry):
+    """Class to contain the information of each entry in
+    tcache_perthread_struct struct.
+
+    Attributes:
+        bk (int): 0 since it is not used.
+    """
+
+    def __init__(self, address, fd, chunks_size):
+        super(TcacheEntry, self).__init__(address, fd, chunks_size=chunks_size)
+
